what
it worked on my machine though

+++ nylium.Networking/DataTypes/Boolean.cs
@@ -24,7 +24,7 @@ namespace nylium.Networking.DataTypes {
         }
 
         public override void Write(Stream stream) {
-            stream.WriteByte(Value == true ? 0x01 : 0x00);
+            stream.WriteByte(Value == true ? (byte) 0x01 : (byte) 0x00);
         }
     }
 }
